Rename InfoWidget folder to InfoWidgets and with this rename the namespace to reflect the content
